Int overflow issue

+++ 08-cython/perf/math_core.pyx.py
@@ -4,9 +4,10 @@ import cython
 from Cython.Shadow import nogil
 
 
-def do_math(start: cython.int = 0, num: cython.int = 10):
-    pos: cython.int = start
-    k_sq: cython.int = 1000 * 1000
+def do_math(start: cython.float = 0, num: cython.float = 10):
+    pos: cython.float = start
+    k_sq: cython.float = 1000 * 1000
+
     with nogil:
         while pos < num:
             pos += 1
